Remove obsolete forceModuleActive setting in EventSystemChecker
This commit addresses a compiler warning in the EventSystemChecker.cs file. The warning was related to the use of the obsolete property 'forceModuleActive' in the StandaloneInputModule. As per the latest Unity documentation, there is no need to set this property, as StandaloneInputModule now functions universally across all platforms. Therefore, the line of code responsible for setting 'forceModuleActive' has been removed to eliminate the warning and maintain codebase compatibility with current best practices.